调整动画编辑形式
