+++ src/components/MiniDashboard/Animation/Animation.module.less
@@ -22,7 +22,3 @@
 .optionlabel {
   background-color: red;
 }
-
-.add {
-  text-align: right;
-}
