--- src/components/MiniDashboard/Animation/AnimationItem/MoveIcon.tsx
@@ -1,14 +1,0 @@
-import * as React from "react"
-
-function MoveIcon() {
-  return (
-    <svg
-      viewBox="0 0 24 24"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path d="M15.5 7h-7C7.673 7 7 6.327 7 5.5c0-.367.134-.719.376-.992L10.873.511c.555-.646 1.695-.65 2.256.003l3.497 3.997c.24.27.374.622.374.989 0 .827-.673 1.5-1.5 1.5zM12 1a.49.49 0 00-.371.166L8.126 5.169A.5.5 0 008.5 6h7a.5.5 0 00.376-.828l-3.503-4.003A.489.489 0 0012 1zM12 24c-.438 0-.849-.188-1.129-.514l-3.497-3.997A1.489 1.489 0 017 18.5c0-.827.673-1.5 1.5-1.5h7c.827 0 1.5.673 1.5 1.5 0 .367-.134.719-.376.992l-3.497 3.997A1.484 1.484 0 0112 24zm-3.5-6a.5.5 0 00-.376.828l3.503 4.003a.512.512 0 00.744.003l3.503-4.003A.5.5 0 0015.5 18zM12 15c-1.654 0-3-1.346-3-3s1.346-3 3-3 3 1.346 3 3-1.346 3-3 3zm0-5c-1.103 0-2 .897-2 2s.897 2 2 2 2-.897 2-2-.897-2-2-2z" />
-    </svg>
-  )
-}
-
-export default MoveIcon
