--- src/components/MiniDashboard/Animation/AnimationItem/index.ts
@@ -1,3 +1,0 @@
-import AnimationItem from "./AnimationItem";
-
-export default AnimationItem;
