--- src/components/MiniDashboard/Animation/AnimationListHoc/index.ts
@@ -1,3 +1,0 @@
-import AnimationListHoc from "./AnimationListHoc";
-
-export default AnimationListHoc;
