+++ src/components/MiniDashboard/NumberInput/NumberInput.tsx
@@ -58,6 +58,7 @@ const QuadrangularSelect: React.FC<Props> = ({
     const onChangeDebounce = useMemo(
         () =>
             throttle((e: number) => {
+                console.log(e);
                 refChange(e);
             }, 500),
         [refChange]
